Add linear channel and related configuration
Linear channels are intended to map selectable RTMP stream
inputs and maintain continuity of output by inserting slate
when RTMP stream sources are unavailable.

At startup a channel will loop a slate asset until the main
RTMP stream comes online and gets mapped to the channel.

All slate and RTMP stream sources must be split into chunks. A
chunk is defined as a colletion of av.Packet instances starting
with a key frame packet followed by the rest of the progress
frames that follow it.

At the moment the only thing that works is the initial slate playback
and that only works until the slate asset loops the first time. This
is due to timestamp mapping issues. Once the metadata packets are
cleaned up and properly correlated into a master time series for the
channel output the RTMP stream generated by the channel should work
much better.